requirements.txt

fastapi==0.112.2
uvicorn==0.30.6
openai==1.42.0
python-dotenv==1.0.1
rapidfuzz==3.9.7
numpy<2
faiss-cpu==1.8.0
packaging>=24.0
pandas==2.2.2
pydantic==2.8.2
twilio==9.3.2
unidecode==1.3.8
sentence-transformers==2.4.1
python-multipart==0.0.20

